+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.17] - 2023-07-15
+
 ### Added
 - Added inbox feed endpoints (getting feed, mark as read, mark as archived etc)